Rewrote the 'dif' function in misc_behr_wind_plots.m some
Mainly focused on the input section, reorganizing the code to make
use of the ask_multichoice general utility function I wrote and
eliminating some questions that are no longer relevant.

Also refactored the parsing section to do away with the overly
complicated method of choosing paths. The new way simply nests switch
and if/elseif statements as needed rather than trying to be overly
clever.

The subsecting of inputs has been automated, so values for xx and yy
will not need to be specified.

Finally plotting WRF data now has the option for wind fields.