New functions and bugfixes to misc_behr_wind_plots.m
MODIFIED FILES:
    Utils/misc_behr_wind_plots.m - multiple changes:
        New subfunction: plot_avg_apriori - averages a priori profiles over an
averaging period. Mainly to check if the use of lonweight in the monthly
averages is biasing the results in some way.

        New subfunction: plot_behr_avg - replicates in many ways the
functionality of no2_column_map_2014 but allows different elements of the
averaging to be turned on and off. Needed to identify why the full and pseudo
retrievals gave different average changes.

        Fix (plot_diff): corrected subscripting of AMF arrays using the new
(on-the-fly) derived xx & yy. Also simplified the color limits to just ensure
that the range is equal above and below 0 for differences and 0 to max for
non-differences.

        New functionality (plot_pseudo_diff_timeser): allowed selection of which
difference in a priori to use, plus a number of new plot options. Also fixed the
function name and added a progress bar.