Some changes to plotting functions to help figure things out for paper.
MODIFIED FILES:
    Utils/misc_behr_wind_plots.m - flipped the order of the daily and hybrid
boxes in the comparison box plot to match the new paper focus on full profiles,
with hybrid as a tool only to separate free trop and BL.
        Also added %delta VCD to wind-cond-v-cld.
        Finally a new function: pix_pos_vs_cld: plots % change in AMF or VCD
against cloud fraction and pixel position in terms of distance and angle from
Atlanta to try to nail down how I want to describe the spatial pattern in the
average change in VCDs.

    WRF_Utils/misc_wrf_wind_plots.m - mainly focused on answering the question
if Atlanta, Birmingham, and Montgomery switch between NOx limited and NOx
suppressed chemistry.
        New subfunction "surface_conc_stats" just plots the min, max, mean, and
median surface conc to get some idea of where things might be on the curve,
since the transition usually happens around 1 ppb.
        Mainly modified the "nonlin-nox" subfunction to bin by VOC reactivity
with OH. Most of the changes support this. There's also a function in the
"Models" repo that can plot Murphy's analytical model.