Updating various functions to use the updated monthly averages (with 1330 LST overpass and incl. 1800 UTC profs)
MODIFIED FILES:
    Emissions/Plotting/fit_var_plotting.m - previous subfunction "pairwise_t"
(which actually did 2-sample t-tests) renamed to t_test and gives option to do
2-sample or paired t-tests.  In the wind effect paper, I decided to keep the
2-sample t-tests because they account for the uncertainty in the measured
values.

    Utils/misc_behr_wind_plots.m - I *think* that I've gotten all the paths
updated to point to the new retrievals (using 1330 and incl. 1800). Also changed
the mag-delvcd subfunction to use a more conservative quadratic sum of errors,
based on Bucsela 13 and references within.

    Utils/subset_BEHR_pixels.m - fixed an issue where if a distance limit from a
geographic coordinate wasn't set, the options structure would not have the
fields center_lon and center_lat. This would cause an error if this structure
was later passed back into the function; these fields are now set to be empty
arrays if the distance limit isn't set.